[X86] Shrink i64 instructions to i32 when the upper 32-bits are known to be zero
(binop x, y) --> (zext (binop (trunc x), (trunc y)))

We can often save a byte in instruction encoding by using the 32-bit variant of the instruction and using the implicit zeroing of the upper 32-bits of the register.

I had hoped to do this in SimplifyDemandedBits to then let further simplifications to occur but we have too many combines that promote bitlogic through extension instructions that I'm not confident will be easy to unravel.