[X86] Handle BSF/BSF "zero-input fall through" behaviour
AMD has always described BSF/BSF as not changing the destination register if the input value was zero, which allows us to support CTTZ/CTLZ zero-input cases by setting the destination to support a NumBits result (BSR is a bit messy as it has to XOR'd to create a CTLZ result).

Currently Intel docs describe this BSF/BSF case as undefined, but its possible they will change this in the near future - this is a pre-emptive patch in case this happens.